terraform/src/clustering_articles/requirements.txt
google-genai==1.15.0
cloud-sql-python-connector==1.18.1
PyMySQL==1.1.1
functions-framework==3.8.3
scikit-learn==1.6.1
numpy==2.2.6
google-cloud-logging==3.12.1
